move endpoint url, test
